Centralize type definitions
